LayoutTests/fast/dom/shadow/shadow-root-activeElement-expected.txt

This tests the activeElement property of a ShadowRoot.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS shadowRoot.activeElement is defined.
PASS shadowRootInside.activeElement is defined.
PASS childInBody.focus();document.activeElement is childInBody
PASS childInBody.focus();shadowRoot.activeElement is null
PASS childInBody.focus();shadowRootInside.activeElement is null
PASS childInTreeOne.focus();document.activeElement is shadowHost
PASS childInTreeOne.focus();shadowRoot.activeElement is childInTreeOne
PASS childInTreeOne.focus();shadowRootInside.activeElement is null
PASS childInTreeTwo.focus();document.activeElement is shadowHost
PASS childInTreeTwo.focus();shadowRoot.activeElement is shadowHostInside
PASS childInTreeTwo.focus();shadowRootInside.activeElement is childInTreeTwo
childInTreeTwo.blur();
PASS document.activeElement is document.body
PASS shadowRoot.activeElement is null
PASS shadowRootInside.activeElement is null
PASS successfullyParsed is true

TEST COMPLETE
